keepalived

+++ backend/setup.sh
@@ -23,12 +23,37 @@ if [[ -z "${AWS_SECRET_ACCESS_KEY}" ]]; then
   exit 1
 fi
 
+if [[ -z "${SRC_IP}" ]]; then
+  echo Missing SRC_IP
+  exit 1
+fi
+
+if [[ -z "${PEER_IP}" ]]; then
+  echo Missing PEER_IP
+  exit 1
+fi
+
+if [[ -z "${EIP}" ]]; then
+  echo Missing EIP
+  exit 1
+fi
+
+if [[ -z "${INSTANCE_ID}" ]]; then
+  echo Missing INSTANCE_ID
+  exit 1
+fi
+
+if [ "$TYPE" != "MASTER" ] && [ "$TYPE" != "BACKUP" ]; then
+  echo TYPE must be MASTER or BACKUP
+  exit 1
+fi
+
 
 # Dependencies
 
 sudo add-apt-repository ppa:jonathonf/python-3.6 -y
 sudo apt update
-sudo apt install linux-headers-$(uname -r) python3.6 make git nginx awscli keepalived -y
+sudo apt install linux-headers-$(uname -r) python3.6 build-essential libssl-dev git nginx awscli -y
 
 sudo apt dist-upgrade -y
 sudo apt autoremove -y
@@ -56,8 +81,9 @@ python manage.py collectstatic
 
 # Systemd
 
-sed -i -e "s/RDS_PASSWORD=/RDS_PASSWORD=${RDS_PASSWORD}/g" /home/ubuntu/infra/backend/env
-sed -i -e "s/RDS_HOSTNAME=/RDS_HOSTNAME=${RDS_HOSTNAME}/g" /home/ubuntu/infra/backend/env
+cp /home/ubuntu/infra/backend/env /home/ubuntu/env
+sed -i -e "s/RDS_PASSWORD=/RDS_PASSWORD=${RDS_PASSWORD}/g" /home/ubuntu/env
+sed -i -e "s/RDS_HOSTNAME=/RDS_HOSTNAME=${RDS_HOSTNAME}/g" /home/ubuntu/env
 
 sudo cp /home/ubuntu/infra/backend/gunicorn.service /etc/systemd/system/gunicorn.service
 sudo systemctl daemon-reload
@@ -87,3 +113,31 @@ echo aws_secret_access_key=$AWS_SECRET_ACCESS_KEY>> ~/.aws/credentials
 
 
 # Keepalived
+
+cd
+wget http://www.keepalived.org/software/keepalived-1.3.9.tar.gz
+tar xf keepalived-1.3.9.tar.gz
+rm -rf keepalived-1.3.9.tar.gz
+cd keepalived-1.3.9
+./configure
+make
+sudo make install
+sudo cp keepalived/etc/init.d/keepalived /etc/init.d/keepalived
+
+sudo cp /home/ubuntu/infra/backend/master.sh /etc/keepalived/master.sh
+sed -i -e "s/EIP/${EIP}/g" /etc/keepalived/master.sh
+sed -i -e "s/INSTANCE_ID/${INSTANCE_ID}/g" /etc/keepalived/master.sh
+
+sudo cp /home/ubuntu/infra/backend/keepalived.conf /etc/keepalived/keepalived.conf
+sed -i -e "s/SRC_IP/${SRC_IP}/g" /etc/keepalived/master.sh
+sed -i -e "s/PEER_IP/${PEER_IP}/g" /etc/keepalived/master.sh
+
+if [ "$TYPE" = "MASTER" ]; then
+  sed -i -e "s/MASTER_OR_BACKUP/MASTER/g" /etc/keepalived/master.sh
+  sed -i -e "s/PRIORITY/101/g" /etc/keepalived/master.sh
+else
+  sed -i -e "s/MASTER_OR_BACKUP/BACKUP/g" /etc/keepalived/master.sh
+  sed -i -e "s/PRIORITY/100/g" /etc/keepalived/master.sh
+fi
+
+sudo service keepalived restart
